Gekko: inline the interpreter hot path and add an x86-64 recompiler

+++ src/gekkodec.h
@@ -435,7 +435,10 @@ namespace Gekko
 
 		size_t numParam;
 		Param param[5];
-		size_t paramBits[5];
+		// Every decoded parameter is an instruction field (a register number, a CR field,
+		// an SPR number, ...), so 16 bits is enough and keeps DecoderInfo small - it is
+		// written and read once per emulated instruction.
+		uint16_t paramBits[5];
 
 		// The value for Immediate parameters is stored here instead of paramBits. I don't know why I did this, it would probably be good to store it in paramBits, but I don't want to redo it anymore.
